@@ -7,15 +7,13 @@
         }
     ],
     "require": {
-        "php": ">=5.5"
+        "php": ">=7.4 || ^8"
     },
     "require-dev": {
-        "phpunit\/phpunit": "<7",
-        "wp-coding-standards\/wpcs": "^0.14.1",
-        "squizlabs\/php_codesniffer": "^3.0.2",
-        "mockery\/mockery": "*",
-        "10up\/wp_mock": "*",
-        "wimg\/php-compatibility": "^8"
+        "phpunit\/phpunit": "^9",
+        "10up\/wp_mock": "^1",
+        "wpdesk\/wp-code-sniffer": "^1",
+        "wpdesk\/phpstan-rules": "^1"
     },
     "autoload": {
         "psr-4": {
@@ -25,7 +23,6 @@
             "src\/Plugin\/WithoutNamespace"
         ]
     },
-    "autoload-dev": {},
     "extra": {
         "text-domain": "wpdesk-wp-builder",
         "translations-folder": "lang",
@@ -39,5 +36,16 @@
         "phpunit-unit-fast": "phpunit --configuration phpunit-unit.xml --no-coverage",
         "phpunit-integration": "phpunit --configuration phpunit-integration.xml --coverage-text --colors=never",
         "phpunit-integration-fast": "phpunit --configuration phpunit-integration.xml --no-coverage"
+    },
+    "archive": {
+        "exclude": [
+            "\/tests"
+        ]
+    },
+    "config": {
+        "allow-plugins": {
+            "dealerdirect\/phpcodesniffer-composer-installer": true,
+            "phpstan\/extension-installer": true
+        }
     }
 }
