@@ -4,7 +4,7 @@ declare( strict_types=1 );
 
 namespace WPDesk\ShopMagic\Api\Controller;
 
-use WPDesk\ShopMagic\Api\Normalizer\WorkflowAutomationNormalizer;
+use WPDesk\ShopMagic\Api\Normalizer\RecipeAutomationNormalizer;
 use WPDesk\ShopMagic\Components\Collections\ArrayCollection;
 use WPDesk\ShopMagic\Recipe\RecipeConverter;
 use WPDesk\ShopMagic\Workflow\Automation\Automation;
@@ -13,16 +13,15 @@ class RecipesController {
 
 	private const DEFAULT_LOCALE = 'en_US';
 
-	/** @var string */
-	private $dir;
+	private string $recipes_dir;
 
-	public function __construct() {
-		$this->dir = __DIR__ . '/../../../config/recipes';
+	public function __construct( string $recipes_dir ) {
+		$this->recipes_dir = $recipes_dir;
 	}
 
-	public function index( WorkflowAutomationNormalizer $normalizer, RecipeConverter $converter ): \WP_REST_Response {
+	public function index( RecipeAutomationNormalizer $normalizer, RecipeConverter $converter ): \WP_REST_Response {
 		$locale       = $this->get_locale();
-		$lang_path    = $this->dir . '/' . $locale;
+		$lang_path    = $this->recipes_dir . '/' . $locale;
 		$recipe_files = (array) scandir( $lang_path );
 		sort( $recipe_files, SORT_NATURAL );
 		$recipes = ( new ArrayCollection( $recipe_files ) )
@@ -46,7 +45,7 @@ class RecipesController {
 
 	private function get_locale(): string {
 		$locale = get_locale();
-		if ( ! file_exists( $this->dir . '/' . $locale ) ) {
+		if ( ! file_exists( $this->recipes_dir . '/' . $locale ) ) {
 			$locale = self::DEFAULT_LOCALE;
 		}
 
