@@ -65,7 +65,7 @@ class FloatType extends AbstractType {
 	 */
 	public function get_conditions(): array {
 		return [
-			self::IS       => __( self::IS, 'shopmagic-for-woocommerce' ),
+			self::IS       => __( 'is', 'shopmagic-for-woocommerce' ),
 			'is_not'       => __( 'is not', 'shopmagic-for-woocommerce' ),
 			'greater_than' => __( 'is greater than', 'shopmagic-for-woocommerce' ),
 			'less_than'    => __( 'is less than', 'shopmagic-for-woocommerce' ),
